Fix several of Visual Studio's code analysis warnings
There are still 134 to go, though I suspect many of them are flase positives.

+++ src/c_console.h
@@ -98,7 +98,7 @@ extern int              consoleheight;
 extern int              consoledirection;
 
 extern int              consolestrings;
-extern int              consolestrings_max;
+extern size_t           consolestrings_max;
 
 extern char             consolecheat[255];
 extern char             consolecheatparm[3];
